Reset dependency
