+++ Src/Color.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef _COLOR_H_
+#define _COLOR_H_
 
 class Color
 {
@@ -15,25 +16,4 @@ public:
 	~Color();
 };
 
-Color::Color()
-{
-}
-
-Color::Color(int x, int y, int z)
-{
-	this->R = x;
-	this->G = y;
-	this->B = z;
-}
-
-Color::Color(int x, int y, int z, int a)
-{
-	this->R = x;
-	this->G = y;
-	this->B = z;
-	this->A = a;
-}
-
-Color::~Color()
-{
-}
+#endif // !_COLOR_H_
