+++ Src/Point.h
@@ -1,29 +1,26 @@
+
+#ifndef _POINT_H_
+#define _POINT_H_
+
 #include <iostream>
 
 using namespace std;
-
-#pragma once
 
 class Point
 {
 public:
 	int x = 0;
 	int y = 0;
-	Point();
-	Point(int X, int Y);
-	~Point();
+	Point() {};
+	Point(int X, int Y)
+	{
+		this->x = X;
+		this->y = Y;
+	}
+	~Point() {}
+	;
 };
 
-Point::Point()
-{
-}
+#endif // !_POINT_H_
 
-Point::Point(int X, int Y)
-{
-	this->x = X;
-	this->y = Y;
-}
 
-Point::~Point()
-{
-}
